script tweak because i killed the bin folder

+++ Create-Release.ps1
@@ -17,5 +17,7 @@ if (Test-Path $binaries) { Remove-Item $binaries -Recurse -Force }
 . $rootFolder\script\Bump-Version.ps1 -Increment Patch
 . $rootFolder\Build-Solution.ps1 -Build $build -Config $config
 
-Remove-Item "$rootFolder\bin\Shimmer.Tests.*.nupkg"
-Remove-Item "$rootFolder\bin\Shimmer.WiXUi.*.nupkg"
+if (Test-Path $binaries) {
+    Remove-Item "$rootFolder\bin\Shimmer.Tests.*.nupkg"
+    Remove-Item "$rootFolder\bin\Shimmer.WiXUi.*.nupkg"
+}
